requirements.txt
python-telegram-bot==21.6
requests==2.32.3
aiohttp==3.10.10